server/requirements.txt
flask
flask-sqlalchemy
flask-admin
flask-cors
google-api-python-client
oauth2client
Flask-Testing
flask-socketio
eventlet
python-intervals